+++ R/globals.R
@@ -23,8 +23,7 @@
 # how to conduct AMR analysis: https://msberends.github.io/AMR/        #
 # ==================================================================== #
 
-globalVariables(c("...length", # for pm_group_split() on R 3.3
-                  ".rowid",
+globalVariables(c(".rowid",
                   "ab",
                   "ab_txt",
                   "angle",